+++ packages/iov-bns-governance/src/index.ts
@@ -1,0 +1,15 @@
+export { CommitteeId, CommitteeIds, committeeIds } from "./committees";
+export { Governor, GovernorOptions } from "./governor";
+export {
+  AccessGuaranteeFundOptions,
+  AddCommitteeMemberOptions,
+  AddValidatorOptions,
+  AmendCommitteeThresholdOptions,
+  AmendCommitteeQuorumOptions,
+  AmendProtocolOptions,
+  DistributeFundsOptions,
+  RemoveCommitteeMemberOptions,
+  RemoveValidatorOptions,
+  ProposalOptions,
+  ProposalType,
+} from "./proposals";